Add cg_reorder_marked_images_to_current. This command moves the marked images to be adjacent to the current image. If current image is '*', and marked images are ABCDE, then:   if direction == -1: 	arrange ABCDE*   elif direction == 1: 	arrange *ABCDE
There are some effects on the ordering of neighboring images not yet resolved yet.
To demonstrate, consider this list, where 12345 are unmarked images, * is the current image,
and ABC are marked images:

  12A3BC4*5

Supposing direction is -1. The algorithm copies everything into a new buffer, as follows:

  I:
    .......*.  (install the current image in its current location)
  II:
    ....ABC*.  (install the marked images, in order, to the left of the current image)
  III:
    1234ABC*5  (install the unmarked images, in order)

And then copies this rearranged buffer back to sxiv's file/thumb arrays.

The case in which the output position of the result is clipped -- for example, consider the above
case if direction were 1, is not currently well handled:

  .......*.
  BC.....*A  (a preferable outcome would be .....*ABC)
  BC12345*A

The case in which the current image is marked is not reliably handled either.
My executive decision on the appropriate outcome on this case is "arrange the total set
of marked images such that the set remains aligned to the correct edge of the current image INDEX."

  To illustrate, suppose  direction=-1 and state = 12A*C3, with * being marked.

  12A*C3

  ^^^

  A*C... ->

  A*C123

  Therefore, the current image's index within the marked images must be determined. After the
  operation, the current image index must be set to the position of A plus
  curimage_index_in_marked.

  The converse case, with direction=1:

  12A*C3

     ^^^

  ...A*C ->

  123A*C

  An alternate option would fix the current image in place, and arrange the images in the order
  they occur in the list.
  That option requires a different sample case:

  12A3*C

  ....*. -> ...A*C -> 123A*C

  (direction is irrelevant in case of the 'arrange set positionally around current image' option)

  For comparison, dir=left and dir=right with current image not marked:

  Left    Right

  12A3*C  12A3*C
  ....*.  ...*..
  ..AC*.  ...*AC
  12AC*3  123*AC